pyproject.toml
[project]
name = "astrix-labs.genesys"
version = "0.1.0"
description = "The intelligence layer for AI memory — scoring, causal inference, lifecycle management, and active forgetting"
authors = [{name = "Genesys Contributors"}]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
keywords = ["memory", "ai", "mcp", "causal-graph", "agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # MCP & API
    "mcp>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",

    # AI / ML
    "openai>=1.30.0",
    "anthropic>=0.30.0",
    "numpy>=1.26.0",

    # SSE
    "sse-starlette>=2.0.0",

    # Utilities
    "python-dotenv>=1.0.0",

    "httpx>=0.27.0",

    # Auth
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
postgres = [
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
]
falkordb = [
    "falkordb>=1.0.0",
    "redis>=5.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.urls]
Homepage = "https://github.com/rishimeka/genesys"
Repository = "https://github.com/rishimeka/genesys"
Issues = "https://github.com/rishimeka/genesys/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.mypy]
python_version = "3.11"
strict = true

[tool.hatch.build.targets.wheel]
packages = ["src/genesys"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
